Source/flask-server/requirements.txt
requests==2.23.0
numpy==1.18.1
flask==1.1.2
Pillow==7.1.2
gunicorn==20.0.4
pytest==5.4.1